refactor: split pipeline/audio.ts into extract.ts, silence.ts, transcribe.ts
- Split 300-line audio.ts into three focused modules behind a barrel:
  - extract.ts: audio extraction via backend or local ffmpeg
  - silence.ts: silence detection + sound bounds computation
  - transcribe.ts: Groq Whisper transcription with caching
- Extract shared extractWordTimings() helper — eliminates duplicate word
  extraction logic between transcribeAudio() and captions renderer
- Fix temp file cleanup in backend path of extractAudio (was leaking temp
  inputs created by resolvePath)
- Add explicit error in detectSilence when cloud backend is active
  (silencedetect requires local ffmpeg stderr parsing)
- Remove unused groqResponseSchema from captions.ts
